+++ src/labelable/__init__.py
@@ -1,3 +1,3 @@
 """Labelable - A general purpose label printing API and UI."""
 
-__version__ = "0.1.1.dev11"
+__version__ = "0.1.1.dev12"